Added CSRF condition on writing functions Changes to API Calls
